+++ R/useful-items.R
@@ -41,10 +41,6 @@ argonInfoCard <- function(value, title = NULL, stat = NULL, stat_icon = NULL,
     if (background_color == "default") text_color <- "text-white" else text_color <- NULL
   else text_color <- NULL
   
-  statCl <- if (!is.null(stat)) {
-    if (stat > 0) "text-success mr-2" else "text-danger mr-2" 
-  }
-  
   infoCardTag <- shiny::tags$div(
     class = cardCl,
     shiny::tags$div(
@@ -64,10 +60,15 @@ argonInfoCard <- function(value, title = NULL, stat = NULL, stat_icon = NULL,
         )
       ),
       # lower part
-      shiny::p(
-        class = "mt-3 mb-0 text-sm",
-        if (!is.null(stat)) shiny::span(class = statCl, stat_icon, stat),
-        shiny::span(class = paste0("h5 ", text_color), description)
+      shiny::fluidRow(
+        class = "mx-2 mt-3 mb-0 text-sm",
+        if (!is.null(stat)) {
+          shiny::tagList(
+            shiny::span(stat_icon, class = "mr-2"),
+            shiny::tagAppendAttributes(shiny::div(stat), class = "mr-2")
+          )
+        },
+        shiny::span(class = paste0("mt-0 h5 ", text_color), description)
       )
     )
   )